requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
prophet>=1.1.4
scikit-learn>=1.3.0
google-api-python-client>=2.100.0
textblob>=0.17.1
python-dotenv>=1.0.0
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
Pillow>=10.0.0
requests>=2.31.0